refactor: consolidate modal files closer to page files
