--- src/modals/index.ts
@@ -1,4 +1,0 @@
-export * from './CategoryModal';
-export * from './CourseModal';
-export * from './GradeModal';
-export * from './SemesterModal';
