+++ src/pages/index.ts
@@ -1,6 +1,6 @@
-export { default as Calculator } from './Calculator';
-export { default as Course } from './Course';
-export { default as Dashboard } from './Dashboard';
+export { default as Calculator } from './Calculator/Calculator';
+export { default as Course } from './Course/Course';
+export { default as Dashboard } from './Dashboard/Dashboard';
 export { default as GradeScale } from './GradeScale';
 export { default as Login } from './Login';
-export { default as Semesters } from './Semesters';
+export { default as Semesters } from './Semesters/Semesters';
